{{cookiecutter.project_slug}}/src/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = "Telegram bot boilerplate with small engine"
authors = ["Carlos Lugones <contact@lugodev.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.5"
python-dotenv = "^0.17.1"
Django = "^3.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
